feat: adiciona os wrappers e a conexão inicial com a blockchain

+++ src/main/java/com/consorcio/projeto_consorcio/cota/CotaRepository.java
@@ -18,5 +18,5 @@ public interface CotaRepository extends JpaRepository<Cota, Long> {
     List<Cota> findByGrupoConsorcioId(Long grupoId);
 
     //conta quantas cotas existem em um grupo
-    long countBygrupoConsorcioId(Long grupoId);
+    int countBygrupoConsorcioId(Long grupoId);
 }
